+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     an empty string
   - Font dictionaries use the new configuration option
     'font.on_missing_unicode_mapping' in their `#to_utf8` method
+* [HexaPDF::Configuration#constantize] to raise error if constant is not found
 
 ### Fixed